% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaker.R
\name{snaker}
\alias{snaker}
\title{snaker}
\usage{
snaker(doc, name = NULL, help = TRUE, version = NULL)
}
\arguments{
\item{doc}{\code{character} vector with command line specification}

\item{name}{Currently not used}

\item{help}{\code{logical} should "-h" or "--help" generate a usage string?}

\item{version}{\code{character}. If supplied the option "-v" generates
the given version number and stops.}
}
\value{
The list of named arguments and parameters.
}
\description{
This function is equivalent to docopt::docopt, except that it can also handle
arguments passed by Snakemake. If the `Snakemake' object exists, then the
arguments are taken from there, otherwise from the command line.
}
