% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaker.R
\name{snaker}
\alias{snaker}
\title{snaker}
\usage{
snaker(doc, args = commandArgs(TRUE), name = NULL, help = TRUE,
  version = NULL, strict = FALSE, strip_names = !strict,
  quoted_args = !strict)
}
\arguments{
\item{doc}{\code{character} vector with command line specification}

\item{args}{\code{character} vector of commandline arguments. 
If \code{args} is missing 
\code{commandArgs(trailingOnly=TRUE)} will be used.}

\item{name}{Currently not used}

\item{help}{\code{logical} should "-h" or "--help" generate a usage string?}

\item{version}{\code{character}. If supplied the option "-v" generates
the given version number and stops.}

\item{strict}{\code{logical} if \code{TRUE} docopt will conform to docopt.py 
in and output (\code{strip_names=FALSE} and \code{quoted_args=FALSE})}

\item{strip_names}{if \code{TRUE} it will remove dashes and angles from the 
resulting names and add these to the resulting list. 
Note that this is different from docopt standard!}

\item{quoted_args}{if \code{TRUE} it will accept quoted arguments. 
Note that this is different from docopt standard!}
}
\value{
The list of named arguments and parameters.
}
\description{
This function is equivalent to docopt::docopt, except that it can also handle
arguments passed by Snakemake. If the `Snakemake' object exists, then the
arguments are taken from there, otherwise from the command line.
}
